auth modified, edit responced for fe

+++ project/views.py
@@ -4,9 +4,10 @@ from django.shortcuts import get_object_or_404
 from rest_framework import status 
 from .serializers import * 
 from .models import * 
-from rest_framework.decorators import api_view 
+from rest_framework.decorators import api_view,parser_classes
 from minio import Minio
 from datetime import datetime
+from django.http import HttpResponseBadRequest,HttpResponseServerError
 from rest_framework import viewsets
 from django.contrib.auth import authenticate, login, logout
 from django.http import HttpResponse
@@ -19,10 +20,12 @@ from django.contrib.auth import get_user_model
 import redis
 import uuid
 from django.conf import settings
+from rest_framework.parsers import MultiPartParser
+
 
 session_storage = redis.StrictRedis(host=settings.REDIS_HOST, port=settings.REDIS_PORT)
 
-user= Users.objects.get(id=2)
+user= Users.objects.get(id=20)
 
 
 @swagger_auto_schema(method='post', request_body=UserSerializer)
@@ -42,6 +45,31 @@ def create(request):
         return Response({'status': 'Success'}, status=200)
     return Response({'status': 'Error', 'error': serializer.errors}, status=status.HTTP_400_BAD_REQUEST)
 
+@api_view(['GET'])
+@permission_classes([IsAuth])
+def user_info(request):
+    print(request.headers.get('Authorization'))
+    try:
+        access_token = request.COOKIES["access_token"]
+        print('access_token', access_token)
+        if session_storage.exists(access_token):
+            email = session_storage.get(access_token).decode('utf-8')
+            user = Users.objects.get(email=email)
+            application = Applications.objects.filter(customer_id=user.id).filter(status=1).first()
+            user_data = {
+                "user_id": user.id,
+                "email": user.email,
+                "is_moderator": user.is_moderator,
+                "current_cart": application.id if application else -1,
+            }
+            print(user_data)
+            return Response(user_data, status=status.HTTP_200_OK)
+        else:
+            return Response({'status': 'Error', 'message': 'Session does not exist'})
+    except:
+        return Response({'status': 'Error', 'message': 'Cookies are not transmitted'})
+
+
 
 # @authentication_classes([])
 # @csrf_exempt
@@ -55,29 +83,34 @@ def login_view(request):
     
     if user is not None:
         random_key = str(uuid.uuid4())
+        application = Applications.objects.filter(customer_id=user.id).filter(status=1).first()
         user_data = {
             "user_id": user.id,
             "email": user.email,
             "is_moderator": user.is_moderator,
-            "session_id": random_key
+            "access_token": random_key,
+            "current_cart": application.id if application else -1,
         }
         session_storage.set(random_key, username)
         response = Response(user_data, status=status.HTTP_201_CREATED)
-        response.set_cookie("session_id", random_key)
+        response.set_cookie("access_token", random_key)
 
         return response
     else:
         return Response({'status': 'Error'}, status=status.HTTP_400_BAD_REQUEST)
 
+@api_view(['Post'])
+@permission_classes([IsAuth])
 def logout_view(request):
-    authorization_header = request.headers.get('Authorization')
-    access_token = authorization_header.split(' ')[1] if authorization_header else None
+    # authorization_header = request.headers.get('Authorization')
+    # access_token = authorization_header.split(' ')[1] if authorization_header else None
+    access_token = request.COOKIES["access_token"]
     if access_token is None:
         message = {"message": "Token is not found in cookie"}
         return Response(message, status=status.HTTP_401_UNAUTHORIZED)
     session_storage.delete(access_token)
     response = Response({'message': 'Logged out successfully'})
-    response.delete_cookie('session_id')
+    response.delete_cookie('access_token')
 
     return response
 
@@ -89,14 +122,44 @@ def logout_view(request):
 
 
 
+@api_view(['Get']) 
+@permission_classes([AllowAny])
+def get_current_cart(request, format=None):
+    print(request.headers)
 
+    authorization_header = request.headers.get('Authorization')
+    access_token = authorization_header.split(' ')[1] if authorization_header else None
+    print(access_token)
 
+    username = session_storage.get(access_token).decode('utf-8')
+    print('lala',username)
+
+    user_id = Users.objects.filter(email=username).values_list('id', flat=True).first()
+
+    application = get_object_or_404(Applications, customer = user_id)
+    if request.method == 'GET':
+        serializer = ApplicationSerializer(application)
+        application_data = serializer.data
+
+        # Получить связанные опции для заявки с полными данными из таблицы Options
+        application_options = Applicationsoptions.objects.filter(application=application)
+        options_data = []
+        for app_option in application_options:
+            option_serializer = OptionSerializer(app_option.option)
+            option_data = option_serializer.data
+            option_data['amount'] = app_option.amount
+            options_data.append(option_data)
+        
+        # Добавить данные об опциях в данные о заявке
+        application_data['options'] = options_data
+        
+        return Response(application_data)
 
 
 
 #GET - получить список всех опций 
-@api_view(['Get']) 
-# @permission_classes([IsModerator])
+@api_view(['GET']) 
+@permission_classes([AllowAny])
 def get_options(request, format=None): 
     
     search_query = request.GET.get('search', '')
@@ -109,21 +172,24 @@ def get_options(request, format=None):
     if category and category != 'Любая категория':
         options = options.filter(category=category)
 
-    serializer = OptionSerializer(options, many=True)
-    
-    application = Applications.objects.filter(customer=user, status=1).first()
-    if application:
-        application_serializer = ApplicationSerializer(application)
-        apps_data = [application_serializer.data]
-    else:
-        apps_data = []
-    
-    response_data = {
-        'apps': apps_data,
-        'options': serializer.data,
-    }
-    
-    return Response(response_data)
+    try:
+        access_token = request.COOKIES["access_token"]
+        username = session_storage.get(access_token).decode('utf-8')
+        print(username)
+        user_ind = Users.objects.filter(email=username).first()
+        application = Applications.objects.filter(customer_id=user_ind.id, status=1).values_list('id', flat=True).first()
+        serializer = OptionSerializer(options, many=True)
+        response_data = {
+            'app_id': application,
+            'options': serializer.data,
+        }
+        return Response(response_data)
+    except:
+        serializer = OptionSerializer(options, many=True)
+        result = {
+            'options': serializer.data,
+        }
+        return Response(result)
  
 #POST - добавить новую опцию 
 @api_view(['Post']) 
@@ -156,10 +222,42 @@ def post_option(request, format=None):
     return Response(serializer.data)
 
 
+@api_view(['POST'])
+@parser_classes([MultiPartParser])
+@permission_classes([AllowAny])
+def postImageToOption(request, pk):
+    if 'file' in request.FILES:
+        file = request.FILES['file']
+        subscription = Options.objects.get(pk=pk, available=True)
+        
+        client = Minio(endpoint="localhost:9000",
+                       access_key='minioadmin',
+                       secret_key='minioadmin',
+                       secure=False)
 
+        bucket_name = 'img'
+        file_name = file.name
+        file_path = "http://localhost:9000/img/" + file_name
+        
+        try:
+            client.put_object(bucket_name, file_name, file, length=file.size, content_type=file.content_type)
+            print("Файл успешно загружен в Minio.")
+            
+            serializer = OptionSerializer(instance=subscription, data={'image': file_path}, partial=True)
+            if serializer.is_valid():
+                serializer.save()
+                return HttpResponse('Image uploaded successfully.')
+            else:
+                return HttpResponseBadRequest('Invalid data.')
+        except Exception as e:
+            print("Ошибка при загрузке файла в Minio:", str(e))
+            return HttpResponseServerError('An error occurred during file upload.')
+
+    return HttpResponseBadRequest('Invalid request.')
 
 #GET - получить одну опцию 
-@api_view(['Get']) 
+@api_view(['GET']) 
+@permission_classes([AllowAny])
 def get_option(request, pk, format=None): 
     option = get_object_or_404(Options, pk=pk) 
     if request.method == 'GET': 
@@ -178,6 +276,7 @@ def put_option(request, pk, format=None):
  
 #PUT - удалить одну опцию 
 @api_view(['Put']) 
+@permission_classes([IsModerator])
 def delete_option(request, pk, format=None):     
     if not Options.objects.filter(pk=pk).exists():
         return Response(f"Опции с таким id не существует!") 
@@ -192,13 +291,24 @@ def delete_option(request, pk, format=None):
  
 #POST - добавить опцию в заявку(если нет открытых заявок, то создать)
 @api_view(['POST'])
+@permission_classes([IsAuth])
 def add_to_application(request, pk):
+    # authorization_header = request.headers.get('Authorization')
+    # access_token = authorization_header.split(' ')[1] if authorization_header else None
+    # print(request.COOKIES["access_token"])
+    access_token = request.COOKIES["access_token"]
+    username = session_storage.get(access_token).decode('utf-8')
+    user = Users.objects.filter(email=username).first()
+    if user is None:
+        print('Не зареган')
+
+
     if not Options.objects.filter(id=pk).exists():
         return Response(f"Опции с таким id не существует!")
 
     option = Options.objects.get(id=pk)
 
-    application = Applications.objects.filter(status=1).last()
+    application = Applications.objects.filter(status=1,customer_id=user.id).last()
 
     if application is None:
         application = Applications.objects.create(customer_id=user.id)
@@ -212,30 +322,27 @@ def add_to_application(request, pk):
         application_option = Applicationsoptions(application=application, option=option, amount=amount)
         application_option.save()
 
-    return Response(status=200)
+    return Response(application.id)
 
 
 
- 
 #GET - получить список всех заявок 
 @api_view(['GET'])
-@permission_classes([AllowAny])
+@permission_classes([IsAuth])
 def get_applications(request, format=None):
-    print(request.headers)
-
-    authorization_header = request.headers.get('Authorization')
-    access_token = authorization_header.split(' ')[1] if authorization_header else None
-    print(access_token)
-
+    
+    access_token = request.COOKIES["access_token"]
     username = session_storage.get(access_token).decode('utf-8')
-    print('lala',username)
-
     user_id = Users.objects.filter(email=username).values_list('id', flat=True).first()
 
     if username is not None and user_id is not None:
-        applications = Applications.objects.filter(customer_id=user_id)
+        user = Users.objects.get(email=username)
+        print(user.is_moderator)
+        if user.is_moderator:
+            applications = Applications.objects
+        else:
+            applications = Applications.objects.filter(customer_id=user_id).exclude(status=2)
         serializer = ApplicationSerializer(applications, many=True)
-        print('lala',serializer.data)
         return Response(serializer.data)
     else:
         return Response("Invalid user", status=status.HTTP_400_BAD_REQUEST)
@@ -263,13 +370,14 @@ def get_application(request, pk, format=None):
         return Response(application_data)
 
 
+
+# @cross_origin(allow_headers=['Content-Type'])
 @api_view(["PUT"])
+@permission_classes([AllowAny]) #need to fix for preflight
 def update_by_user(request, pk):
     if not Applications.objects.filter(pk=pk).exists():
         return Response(f"Заявки с таким id не существует!")
-
     request_status = request.data["status"]
-
     if int(request.data["status"]) not in [2, 3]:
         return Response(status=status.HTTP_405_METHOD_NOT_ALLOWED)
 
@@ -284,10 +392,19 @@ def update_by_user(request, pk):
     application.save()
 
     serializer = ApplicationSerializer(application, many=False)
-    return Response(serializer.data)
+    response = Response(serializer.data)
+    
+    return response
 
+
+@swagger_auto_schema(method='put',request_body=ApplicationSerializer)
 @api_view(["PUT"])
+@permission_classes([IsModerator])
 def update_by_admin(request, pk):
+    access_token = request.COOKIES["access_token"]
+    modername = session_storage.get(access_token).decode('utf-8')
+    user = Users.objects.filter(email=modername).first()
+    
     if not Applications.objects.filter(pk=pk).exists():
         return Response(f"Заявки с таким id не существует!")
 
@@ -298,17 +415,19 @@ def update_by_admin(request, pk):
 
     application = Applications.objects.get(pk=pk)
     if int(request.data["status"]) in [4]:
-        application.formed_at=timezone.now()
+        application.completed_at=timezone.now()
     # app_status = application.status
 
     # if app_status == 5:
     #     return Response("Статус изменить нельзя")
-
+    application.moderator_id=user.id
     application.status = request_status
     application.save()
 
     serializer = ApplicationSerializer(application, many=False)
-    return Response(serializer.data)
+    response = Response(serializer.data)
+    response.setHeader("Access-Control-Allow-Methods", "PUT")
+    return response
 
 
 #DELETE - удалить одну заявку
@@ -326,6 +445,7 @@ def delete_application(request, pk, format=None):
 
 #DELETE - удалить конкретную опцию из конкретной заявки
 @api_view(["DELETE"])
+@permission_classes([AllowAny])
 def delete_option_from_application(request, application_id, option_id):
     if not Applications.objects.filter(pk=application_id).exists():
         return Response("Заявки с таким id не существует", status=status.HTTP_404_NOT_FOUND)
@@ -336,10 +456,14 @@ def delete_option_from_application(request, application_id, option_id):
     application = Applications.objects.get(pk=application_id)
     option = Options.objects.get(pk=option_id)
 
+    application_subscription = get_object_or_404(Applicationsoptions, application=application, option=option)
+    if application_subscription is None:
+        print('kakakakakak')
+        return Response("Заявка не найдена", status=404)
     application.applicationsoptions_set.filter(option=option).delete()
     application.save()
 
-    return Response("Опция успешно удалена из заявки", status=status.HTTP_204_NO_CONTENT)
+    return Response("Опция успешно удалена из заявки", status=200)
 
 
 #PUT - изменить кол-во конкретной опции в заявке
@@ -361,3 +485,8 @@ def update_option_amount(request, application_id, option_id):
 
     # else:
     #     return Response("Неверные данные для обновления amount", status=status.HTTP_400_BAD_REQUEST)
+
+
+
+
+
